code update

+++ src/main/java/in/ashokit/repo/ProductCategoryRepo.java
@@ -3,7 +3,7 @@ package in.ashokit.repo;
 import in.ashokit.entity.ProductCategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductCategoryRepo extends JpaRepository<ProductCategoryEntity, Long> {
+public interface ProductCategoryRepo extends JpaRepository<ProductCategoryEntity, Integer> {
 
     public ProductCategoryEntity findByCategoryNameIgnoreCase(String categoryName);
 }
